+++ tests/integration/test_model.py
@@ -97,7 +97,7 @@ async def test_deploy_local_charm(event_loop):
     async with base.CleanModel() as model:
         await model.deploy(str(charm_path))
         assert 'charm' in model.applications
-        await model.wait_for_idle(wait_for_active=True)
+        await model.wait_for_idle(status="active")
         assert model.units['charm/0'].workload_status == 'active'
 
 
@@ -113,7 +113,7 @@ async def test_wait_local_charm_blocked(event_loop):
         assert 'charm' in model.applications
         await model.wait_for_idle()
         with pytest.raises(JujuUnitError):
-            await model.wait_for_idle(wait_for_active=True,
+            await model.wait_for_idle(status="active",
                                       raise_on_blocked=True,
                                       timeout=30)
 
@@ -130,7 +130,7 @@ async def test_wait_local_charm_waiting_timeout(event_loop):
         assert 'charm' in model.applications
         await model.wait_for_idle()
         with pytest.raises(asyncio.TimeoutError):
-            await model.wait_for_idle(wait_for_active=True, timeout=30)
+            await model.wait_for_idle(status="active", timeout=30)
 
 
 @base.bootstrapped
@@ -327,27 +327,30 @@ async def test_add_manual_machine_ssh(event_loop):
         # and pylxd's container.execute seems to be broken and fails and/or
         # hangs trying to properly check if the service is up.
         time.sleep(5)
-
+        spec = 'ssh:{}@{}:{}'.format(
+            test_user,
+            host['address'],
+            private_key_path,
+        )
+        err = None
         for attempt in range(1, 4):
             try:
                 # add a new manual machine
-                machine1 = await model.add_machine(spec='ssh:{}@{}:{}'.format(
-                    test_user,
-                    host['address'],
-                    private_key_path,
-                ))
-            except paramiko.ssh_exception.NoValidConnectionsError:
+                machine1 = await model.add_machine(spec=spec)
+            except (paramiko.ssh_exception.NoValidConnectionsError, paramiko.ssh_exception.AuthenticationException) as e:
                 # retry the ssh connection a few times if it fails
+                err = e
                 time.sleep(attempt * 5)
             else:
+                # try part finished without exception, breaking
                 break
 
-            assert len(model.machines) == 1
+        assert len(model.machines) == 1, 'Unable to add_machine in %s attempts with spec : %s -- exception was %s' % (attempt, spec, err)
 
-            res = await machine1.destroy(force=True)
+        res = await machine1.destroy(force=True)
 
-            assert res is None
-            assert len(model.machines) == 0
+        assert res is None, 'Bad teardown, res is : %s' % res
+        assert len(model.machines) == 0
 
         container.stop(wait=True)
         container.delete(wait=True)
@@ -441,25 +444,29 @@ async def test_add_manual_machine_ssh_root(event_loop):
         # and pylxd's container.execute seems to be broken and fails and/or
         # hangs trying to properly check if the service is up.
         time.sleep(5)
-
+        spec = 'ssh:{}@{}:{}'.format(
+            "root",
+            host['address'],
+            private_key_path,
+        )
+        err = None
         for attempt in range(1, 4):
             try:
                 # add a new manual machine
-                machine1 = await model.add_machine(spec='ssh:{}@{}:{}'.format(
-                    "root",
-                    host['address'],
-                    private_key_path,
-                ))
-            except (paramiko.ssh_exception.NoValidConnectionsError, paramiko.ssh_exception.AuthenticationException):
+                machine1 = await model.add_machine(spec=spec)
+            except (paramiko.ssh_exception.NoValidConnectionsError, paramiko.ssh_exception.AuthenticationException) as e:
+                # if we get an exception, try again
+                err = e
                 time.sleep(attempt * 5)
             else:
+                # try part finished without exception
                 break
 
-        assert len(model.machines) == 1
+        assert len(model.machines) == 1, 'Unable to add_machine in %s attempts with spec : %s -- exception was %s' % (attempt, spec, err)
 
         res = await machine1.destroy(force=True)
 
-        assert res is None
+        assert res is None, 'Bad teardown, res is : %s' % res
         assert len(model.machines) == 0
 
         container.stop(wait=True)
@@ -566,6 +573,25 @@ async def test_store_resources_charm(event_loop):
         # ghost will go in to blocked (or error, for older
         # charm revs) if the resource is missing
         assert ghost.units[0].workload_status == 'active'
+
+
+@base.bootstrapped
+@pytest.mark.asyncio
+async def test_local_oci_image_resource_charm(event_loop):
+    tests_dir = Path(__file__).absolute().parent
+    charm_path = tests_dir / 'oci-image-charm'
+    async with base.CleanModel() as model:
+        resources = {"oci-image": "ubuntu/latest"}
+        charm = await model.deploy(str(charm_path), resources=resources)
+        assert 'oci-image-charm' in model.applications
+        terminal_statuses = ('active', 'error', 'blocked')
+        await model.block_until(
+            lambda: (
+                len(charm.units) > 0 and
+                charm.units[0].workload_status in terminal_statuses),
+            timeout=120,
+        )
+        assert charm.units[0].workload_status == 'active'
 
 
 @base.bootstrapped